Use inheritance to extract Contact class
